requirements.txt
pillow
opencv-python
boto3
tqdm